requirements.txt
absl-py==0.11.0
attrs==19.3.0
colorama==0.4.3
cvxpy==1.1.15
datasets==1.8.0
datashader==0.12.0
flask==2.3.2
h5py==2.10.0
holoviews==1.14.1
huggingface-hub==0.0.10
imagenet-x==0.0.3
importlib-metadata==1.5.0
matplotlib==3.1.2
mock==3.0.5
numexpr==2.8.3
numpy==1.19.5
psutil==5.5.1
PyQt5==5.14.1
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.11.0
sentencepiece==0.1.96
sympy==1.5.1
tensorflow==2.4.0
tensorflow-addons==0.11.2
tensorflow-datasets==4.6.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
tensorflow-probability==0.12.1
tensorflow-text==2.4.3
tfds-nightly==3.2.1.dev202007280106
tokenizers==0.10.1
torch==1.11.0
transformers==4.5.0
umap-learn==0.5.1
wolframclient==1.1.7
wrapt==1.12.1
wrapt-timeout-decorator==1.3.12.2
xarray==0.16.2
git+https://github.com/mmatena/npeff_ref.git
